pyproject.toml
[project]
name = "ai-agent-platform"
version = "0.1.0"
description = "A platform that enables users to create, manage, and interact with AI agents via text and voice."
readme = "README.md"
requires-python = ">=3.10.0, <4.0.0"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "httpx>=0.28.1",
    "openai-agents>=0.6.4",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "sqlmodel>=0.0.31",
]
